fix: Handle nil gRPC client in Site Agent workflows (#235)

+++ site-workflow/pkg/activity/dpuextensionservice.go
@@ -132,6 +132,9 @@ func (mdes *ManageDpuExtensionService) CreateDpuExtensionServiceOnSite(ctx conte
 
 	// Call Site Controller gRPC endpoint
 	carbideClient := mdes.CarbideAtomicClient.GetClient()
+	if carbideClient == nil {
+		return nil, cclient.ErrClientNotConnected
+	}
 	forgeClient := carbideClient.Carbide()
 
 	createdDpuExtensionService, err := forgeClient.CreateDpuExtensionService(ctx, request)
@@ -166,6 +169,9 @@ func (mdes *ManageDpuExtensionService) UpdateDpuExtensionServiceOnSite(ctx conte
 
 	// Call Site Controller gRPC endpoint
 	carbideClient := mdes.CarbideAtomicClient.GetClient()
+	if carbideClient == nil {
+		return nil, cclient.ErrClientNotConnected
+	}
 	forgeClient := carbideClient.Carbide()
 
 	updatedDpuExtensionService, err := forgeClient.UpdateDpuExtensionService(ctx, request)
@@ -200,6 +206,9 @@ func (mdes *ManageDpuExtensionService) DeleteDpuExtensionServiceOnSite(ctx conte
 
 	// Call Site Controller gRPC endpoint
 	carbideClient := mdes.CarbideAtomicClient.GetClient()
+	if carbideClient == nil {
+		return cclient.ErrClientNotConnected
+	}
 	forgeClient := carbideClient.Carbide()
 
 	_, err = forgeClient.DeleteDpuExtensionService(ctx, request)
@@ -234,6 +243,9 @@ func (mdes *ManageDpuExtensionService) GetDpuExtensionServiceVersionsInfoOnSite(
 
 	// Call Site Controller gRPC endpoint
 	carbideClient := mdes.CarbideAtomicClient.GetClient()
+	if carbideClient == nil {
+		return nil, cclient.ErrClientNotConnected
+	}
 	forgeClient := carbideClient.Carbide()
 
 	versionInfos, err := forgeClient.GetDpuExtensionServiceVersionsInfo(ctx, request)
