+++ site-workflow/pkg/activity/ibpartition.go
@@ -150,6 +150,9 @@ func (mibp *ManageInfiniBandPartition) CreateInfiniBandPartitionOnSite(ctx conte
 
 	// Call Site Controller gRPC endpoint
 	carbideClient := mibp.CarbideAtomicClient.GetClient()
+	if carbideClient == nil {
+		return client.ErrClientNotConnected
+	}
 	forgeClient := carbideClient.Carbide()
 
 	// Call Forge gRPC endpoint
@@ -185,6 +188,9 @@ func (mipb *ManageInfiniBandPartition) DeleteInfiniBandPartitionOnSite(ctx conte
 
 	// Call Site Controller gRPC endpoint
 	carbideClient := mipb.CarbideAtomicClient.GetClient()
+	if carbideClient == nil {
+		return client.ErrClientNotConnected
+	}
 	forgeClient := carbideClient.Carbide()
 
 	_, err = forgeClient.DeleteIBPartition(ctx, request)
