+++ site-workflow/pkg/activity/networksecuritygroup.go
@@ -65,6 +65,9 @@ func (mm *ManageNetworkSecurityGroup) CreateNetworkSecurityGroupOnSite(ctx conte
 
 	// Call Site Controller gRPC endpoint
 	carbideClient := mm.CarbideAtomicClient.GetClient()
+	if carbideClient == nil {
+		return cClient.ErrClientNotConnected
+	}
 	forgeClient := carbideClient.Carbide()
 
 	_, err = forgeClient.CreateNetworkSecurityGroup(ctx, request)
@@ -102,6 +105,9 @@ func (mm *ManageNetworkSecurityGroup) UpdateNetworkSecurityGroupOnSite(ctx conte
 
 	// Call Site Controller gRPC endpoint
 	carbideClient := mm.CarbideAtomicClient.GetClient()
+	if carbideClient == nil {
+		return cClient.ErrClientNotConnected
+	}
 	forgeClient := carbideClient.Carbide()
 
 	_, err = forgeClient.UpdateNetworkSecurityGroup(ctx, request)
@@ -139,6 +145,9 @@ func (mm *ManageNetworkSecurityGroup) DeleteNetworkSecurityGroupOnSite(ctx conte
 
 	// Call Site Controller gRPC endpoint
 	carbideClient := mm.CarbideAtomicClient.GetClient()
+	if carbideClient == nil {
+		return cClient.ErrClientNotConnected
+	}
 	forgeClient := carbideClient.Carbide()
 
 	_, err = forgeClient.DeleteNetworkSecurityGroup(ctx, request)
